+++ src/Clusterer.h
@@ -18,7 +18,7 @@ public:
 	// Analyzing and storing the hits
         bool AnalyzeHits(double srsTimestamp, uint8_t fecId, 
         uint8_t vmmId, uint16_t chNo, uint16_t bcid, uint16_t tdc, 
-        uint16_t adc, bool overThresholdFlag, float chipTime);
+        uint16_t adc, bool overThresholdFlag, double chipTime);
 
         // Analyzing and storing the clusters in plane 0 and 1
         void AnalyzeClustersPlane(std::pair<uint8_t, uint8_t> dp);
